requirements.txt
python-dotenv
requests
icecream
openai
langchain
langchain-openai
transformers
qdrant-client
tiktoken